document generator ui, wikipedia api fix, google api fix, generator fixed data exclusion

+++ wikipedia_api.py
@@ -1,4 +1,4 @@
-import wikipedia
+import wikipediaapi
 import util
 
 
@@ -8,6 +8,12 @@ __author__ = 'Ziga Vucko'
 def run(query, text_dir):
     # Run the query to the Wikipedia API.
     print(util.timestamp() + ' Calling the Wikipedia API, downloading the content of the given term and saving it to cache')
+
+    wikipedia = wikipediaapi.Wikipedia('en')
+
     page = wikipedia.page(query)
+    if not page.exists():
+        raise util.Error('Wikipedia page for keyword "' + query + '" does not exist.')
+
     with open(text_dir + '/wikipedia.txt', 'w') as file:
-        file.write(page.content)
+        file.write(page.text)
